Split some actions Pt 3 (#3795)
moving Annotation actions AnnotationActions
