+++ client/app/reader/SideBarComments.jsx
@@ -7,8 +7,9 @@ import { plusIcon } from '../components/RenderFunctions';
 import Button from '../components/Button';
 import _ from 'lodash';
 import { INTERACTION_TYPES } from './analytics';
-import { updateNewAnnotationContent, createAnnotation, stopPlacingAnnotation,
-  startPlacingAnnotation } from '../reader/actions';
+import { createAnnotation, stopPlacingAnnotation,
+  startPlacingAnnotation, updateNewAnnotationContent
+} from '../reader/PdfViewer/AnnotationActions';
 
 class SideBarComments extends PureComponent {
   handleAddClick = (event) => {
