+++ client/app/reader/actions.js
@@ -3,9 +3,9 @@
 import * as Constants from './constants';
 import _ from 'lodash';
 import ApiUtil from '../util/ApiUtil';
-import uuid from 'uuid';
 import { CATEGORIES, ENDPOINT_NAMES } from './analytics';
 import { createSearchAction } from 'redux-search';
+import { selectAnnotation } from '../reader/PdfViewer/AnnotationActions';
 
 export const collectAllTags = (documents) => ({
   type: Constants.COLLECT_ALL_TAGS_FOR_OPTIONS,
@@ -15,19 +15,6 @@ export const collectAllTags = (documents) => ({
 export const onScrollToComment = (scrollToComment) => ({
   type: Constants.SCROLL_TO_COMMENT,
   payload: { scrollToComment }
-});
-
-export const startEditAnnotation = (annotationId) => ({
-  type: Constants.START_EDIT_ANNOTATION,
-  payload: {
-    annotationId
-  },
-  meta: {
-    analytics: {
-      category: CATEGORIES.VIEW_DOCUMENT_PAGE,
-      action: 'start-edit-annotation'
-    }
-  }
 });
 
 export const openAnnotationDeleteModal = (annotationId, analyticsLabel) => ({
@@ -43,6 +30,7 @@ export const openAnnotationDeleteModal = (annotationId, analyticsLabel) => ({
     }
   }
 });
+
 export const closeAnnotationDeleteModal = () => ({
   type: Constants.CLOSE_ANNOTATION_DELETE_MODAL,
   meta: {
@@ -50,116 +38,6 @@ export const closeAnnotationDeleteModal = () => ({
       category: CATEGORIES.VIEW_DOCUMENT_PAGE,
       action: 'close-annotation-delete-modal'
     }
-  }
-});
-export const selectAnnotation = (annotationId) => ({
-  type: Constants.SELECT_ANNOTATION,
-  payload: {
-    annotationId
-  },
-  meta: {
-    analytics: {
-      category: CATEGORIES.VIEW_DOCUMENT_PAGE,
-      action: 'select-annotation'
-    }
-  }
-});
-
-export const deleteAnnotation = (docId, annotationId) =>
-  (dispatch) => {
-    dispatch({
-      type: Constants.REQUEST_DELETE_ANNOTATION,
-      payload: {
-        annotationId
-      },
-      meta: {
-        analytics: {
-          category: CATEGORIES.VIEW_DOCUMENT_PAGE,
-          action: 'request-delete-annotation'
-        }
-      }
-    });
-
-    ApiUtil.delete(`/document/${docId}/annotation/${annotationId}`, {}, ENDPOINT_NAMES.ANNOTATION).
-      then(
-        () => dispatch({
-          type: Constants.REQUEST_DELETE_ANNOTATION_SUCCESS,
-          payload: {
-            annotationId
-          }
-        }),
-        () => dispatch({
-          type: Constants.REQUEST_DELETE_ANNOTATION_FAILURE,
-          payload: {
-            annotationId
-          }
-        })
-      );
-  };
-
-export const requestMoveAnnotation = (annotation) => (dispatch) => {
-  dispatch({
-    type: Constants.REQUEST_MOVE_ANNOTATION,
-    payload: {
-      annotation
-    },
-    meta: {
-      analytics: {
-        category: CATEGORIES.VIEW_DOCUMENT_PAGE,
-        action: 'request-move-annotation'
-      }
-    }
-  });
-
-  const data = ApiUtil.convertToSnakeCase({ annotation });
-
-  ApiUtil.patch(`/document/${annotation.documentId}/annotation/${annotation.id}`, { data }, ENDPOINT_NAMES.ANNOTATION).
-    then(
-      () => dispatch({
-        type: Constants.REQUEST_MOVE_ANNOTATION_SUCCESS,
-        payload: {
-          annotationId: annotation.id
-        }
-      }),
-      () => dispatch({
-        type: Constants.REQUEST_MOVE_ANNOTATION_FAILURE,
-        payload: {
-          annotationId: annotation.id
-        }
-      })
-    );
-};
-
-export const cancelEditAnnotation = (annotationId) => ({
-  type: Constants.CANCEL_EDIT_ANNOTATION,
-  payload: {
-    annotationId
-  },
-  meta: {
-    analytics: {
-      category: CATEGORIES.VIEW_DOCUMENT_PAGE,
-      action: 'cancel-edit-annotation'
-    }
-  }
-});
-export const updateAnnotationContent = (content, annotationId) => ({
-  type: Constants.UPDATE_ANNOTATION_CONTENT,
-  payload: {
-    annotationId,
-    content
-  },
-  meta: {
-    analytics: {
-      category: CATEGORIES.VIEW_DOCUMENT_PAGE,
-      action: 'edit-annotation-content-locally',
-      debounceMs: 500
-    }
-  }
-});
-export const updateNewAnnotationContent = (content) => ({
-  type: Constants.UPDATE_NEW_ANNOTATION_CONTENT,
-  payload: {
-    content
   }
 });
 
@@ -180,127 +58,6 @@ export const jumpToPage = (pageNumber, docId) => ({
 export const resetJumpToPage = () => ({
   type: Constants.RESET_JUMP_TO_PAGE
 });
-
-export const requestEditAnnotation = (annotation) => (dispatch) => {
-  // If the user removed all text content in the annotation, ask them if they're
-  // intending to delete it.
-  if (!annotation.comment) {
-    dispatch(openAnnotationDeleteModal(annotation.id, 'open-by-deleting-all-annotation-content'));
-
-    return;
-  }
-
-  dispatch({
-    type: Constants.REQUEST_EDIT_ANNOTATION,
-    payload: {
-      annotationId: annotation.id
-    },
-    meta: {
-      analytics: {
-        category: CATEGORIES.VIEW_DOCUMENT_PAGE,
-        action: 'request-edit-annotation'
-      }
-    }
-  });
-
-  const data = ApiUtil.convertToSnakeCase({ annotation });
-
-  ApiUtil.patch(`/document/${annotation.documentId}/annotation/${annotation.id}`, { data }, ENDPOINT_NAMES.ANNOTATION).
-    then(
-      () => dispatch({
-        type: Constants.REQUEST_EDIT_ANNOTATION_SUCCESS,
-        payload: {
-          annotationId: annotation.id
-        }
-      }),
-      () => dispatch({
-        type: Constants.REQUEST_EDIT_ANNOTATION_FAILURE,
-        payload: {
-          annotationId: annotation.id
-        }
-      })
-    );
-};
-
-export const startPlacingAnnotation = (interactionType) => ({
-  type: Constants.START_PLACING_ANNOTATION,
-  meta: {
-    analytics: {
-      category: CATEGORIES.VIEW_DOCUMENT_PAGE,
-      action: 'start-placing-annotation',
-      label: interactionType
-    }
-  }
-});
-
-export const showPlaceAnnotationIcon = (pageIndex, pageCoords) => ({
-  type: Constants.SHOW_PLACE_ANNOTATION_ICON,
-  payload: {
-    pageIndex,
-    pageCoords
-  }
-});
-
-export const placeAnnotation = (pageNumber, coordinates, documentId) => ({
-  type: Constants.PLACE_ANNOTATION,
-  payload: {
-    page: pageNumber,
-    x: coordinates.xPosition,
-    y: coordinates.yPosition,
-    documentId
-  }
-});
-
-export const stopPlacingAnnotation = (interactionType) => ({
-  type: Constants.STOP_PLACING_ANNOTATION,
-  meta: {
-    analytics: {
-      category: CATEGORIES.VIEW_DOCUMENT_PAGE,
-      action: 'stop-placing-annotation',
-      label: interactionType
-    }
-  }
-});
-
-export const createAnnotation = (annotation) => (dispatch) => {
-  const temporaryId = uuid.v4();
-
-  dispatch({
-    type: Constants.REQUEST_CREATE_ANNOTATION,
-    payload: {
-      annotation: {
-        ...annotation,
-        id: temporaryId
-      }
-    }
-  });
-
-  const data = ApiUtil.convertToSnakeCase({ annotation });
-
-  ApiUtil.post(`/document/${annotation.documentId}/annotation`, { data }, ENDPOINT_NAMES.ANNOTATION).
-    then(
-      (response) => {
-        const responseObject = JSON.parse(response.text);
-
-        dispatch({
-          type: Constants.REQUEST_CREATE_ANNOTATION_SUCCESS,
-          payload: {
-            annotation: {
-              ...annotation,
-              ...responseObject
-            },
-            annotationTemporaryId: temporaryId
-          }
-        });
-      },
-      () => dispatch({
-        type: Constants.REQUEST_CREATE_ANNOTATION_FAILURE,
-        payload: {
-          annotationTemporaryId: temporaryId
-        }
-      })
-    );
-};
 
 export const handleSelectCommentIcon = (comment) => (dispatch) => {
   // Normally, we would not want to fire two actions here.
